make the epson preprocessor the default
While the original plan was to potentially support more printer types,
it's not likely to happen in this universe. The other useless preprocessors
(crlf, simple) are not ging to be used ever and if someone uses dotprint,
it's for the epson preprocessor. So no need forcing users to always
add "-P epson".